+Goal routes are organised for manipulation for own users

+++ controllers/goalController.js
@@ -27,6 +27,8 @@ export const createGoal = asyncHandler(async (req, res) => {
 });
 
 export const updateGoal = asyncHandler(async (req, res) => {
+  const goal = await Goal.findById(req.params.goalId);
+
   if (!req.body.text) {
     res.status(400);
     throw new Error("Please add a text field");
@@ -34,9 +36,14 @@ export const updateGoal = asyncHandler(async (req, res) => {
 
   const { text } = req.body;
 
-  const goal = await Goal.findByIdAndUpdate(req.params.goalId, { text });
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("You have no such this goal");
+  }
 
-  if (!goal) {
+  const updatedGoal = await Goal.findByIdAndUpdate(req.params.goalId, { text });
+
+  if (!updatedGoal) {
     res.status(400);
     throw new Error("Goal not found");
   }
@@ -45,7 +52,14 @@ export const updateGoal = asyncHandler(async (req, res) => {
 });
 
 export const deleteGoal = asyncHandler(async (req, res) => {
-  const deletedGoal = await Goal.findByIdAndRemove(req.params.goalId);
+  const goal = await Goal.findById(req.params.goalId);
 
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("You have no such this goal");
+  }
+
+  const deletedGoal = await Goal.findByIdAndRemove(req.params.goalId);
+  
   res.status(200).json(deletedGoal);
 });
